+++ docs/history.md
@@ -18,7 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ***Added:***
 
 - Add context formatting functionality i.e. the ability to insert values into configuration like environment variables and command line arguments
-- Add `--show-headers` option to the `env run` command to always display headers, even for single environments
+- Any verbosity for command execution will now always display headers, even for single environments
+- Every executed command is now displayed when running multiple commands or when verbosity is enabled
 - Similar to `make`, ignore the exit code of executed commands that start with `-` (a hyphen)
 - Update project metadata to reflect the adoption by PyPA